requirements.txt
psycopg2-binary
requests
pandas
astropy
flask

pytest
pytest-mock
pytest-iterassert